+++ src/jace/util/__init__.py
@@ -9,6 +9,7 @@
 
 from __future__ import annotations
 
+from .definitions import FORBIDDEN_SDFG_VAR_NAMES, VALID_SDFG_OBJ_NAME, VALID_SDFG_VAR_NAME
 from .jax_helper import (
     JaCeVar,
     get_jax_var_dtype,
@@ -18,7 +19,6 @@ from .jax_helper import (
     propose_jax_name,
     translate_dtype,
 )
-from .misc import FORBIDDEN_SDFG_VAR_NAMES, VALID_SDFG_OBJ_NAME, VALID_SDFG_VAR_NAME
 from .traits import (
     is_array,
     is_drop_var,
